Move lua bin parser code to ##bin formats
